add feature service forms
