+++ src/app/kidsSafetyPre/page.jsx
@@ -1,5 +1,5 @@
 'use client'
-import KidsSafetyPreview from "@/components/servicesContent/KidsSafetyPreview";
+import KidsSafetyPreview from "@/components/servicesPages/servicesLayout/contentTabs/KidsSafetyPreview";
 
 export default function kidsSafetyPre() {
   return <KidsSafetyPreview />
