+++ src/context/qrCodeDesignContext/DesignProvider.jsx
@@ -5,7 +5,7 @@
 // import { DesignContext } from "./DesignContext";
 
 // const DesignProvider = ({ children }) => {
- 
+
 
 //         const [selectedQRShape, setSelectedQRShape] = useState(null);
 //         const [selectedLogo, setSelectedLogo] = useState(null);
